+++ user_guide/_generated/06/74-undo-to.sh
@@ -1,1 +1,0 @@
-tether undo --to cda6e861821c